+++ mobile/test/fixtures/asset.stub.dart
@@ -1,10 +1,11 @@
+import 'package:immich_mobile/domain/models/asset/base_asset.model.dart';
 import 'package:immich_mobile/domain/models/exif.model.dart';
-import 'package:immich_mobile/entities/asset.entity.dart';
+import 'package:immich_mobile/entities/asset.entity.dart' as old;
 
 final class AssetStub {
   const AssetStub._();
 
-  static final image1 = Asset(
+  static final image1 = old.Asset(
     checksum: "image1-checksum",
     localId: "image1",
     remoteId: 'image1-remote',
@@ -13,7 +14,7 @@ final class AssetStub {
     fileModifiedAt: DateTime(2020),
     updatedAt: DateTime.now(),
     durationInSeconds: 0,
-    type: AssetType.image,
+    type: old.AssetType.image,
     fileName: "image1.jpg",
     isFavorite: true,
     isArchived: false,
@@ -21,7 +22,7 @@ final class AssetStub {
     exifInfo: const ExifInfo(isFlipped: false),
   );
 
-  static final image2 = Asset(
+  static final image2 = old.Asset(
     checksum: "image2-checksum",
     localId: "image2",
     remoteId: 'image2-remote',
@@ -30,7 +31,7 @@ final class AssetStub {
     fileModifiedAt: DateTime(2010),
     updatedAt: DateTime.now(),
     durationInSeconds: 60,
-    type: AssetType.video,
+    type: old.AssetType.video,
     fileName: "image2.jpg",
     isFavorite: false,
     isArchived: false,
@@ -38,7 +39,7 @@ final class AssetStub {
     exifInfo: const ExifInfo(isFlipped: true),
   );
 
-  static final image3 = Asset(
+  static final image3 = old.Asset(
     checksum: "image3-checksum",
     localId: "image3",
     ownerId: 1,
@@ -46,10 +47,30 @@ final class AssetStub {
     fileModifiedAt: DateTime(2025),
     updatedAt: DateTime.now(),
     durationInSeconds: 60,
-    type: AssetType.image,
+    type: old.AssetType.image,
     fileName: "image3.jpg",
     isFavorite: true,
     isArchived: false,
     isTrashed: false,
+  );
+}
+
+abstract final class LocalAssetStub {
+  const LocalAssetStub._();
+
+  static final image1 = LocalAsset(
+    id: "image1",
+    name: "image1.jpg",
+    type: AssetType.image,
+    createdAt: DateTime(2025),
+    updatedAt: DateTime(2025, 2),
+  );
+
+  static final image2 = LocalAsset(
+    id: "image2",
+    name: "image2.jpg",
+    type: AssetType.image,
+    createdAt: DateTime(2000),
+    updatedAt: DateTime(20021),
   );
 }
