+++ mobile/test/infrastructure/repository.mock.dart
@@ -1,5 +1,8 @@
 import 'package:immich_mobile/domain/interfaces/device_asset.interface.dart';
+import 'package:immich_mobile/domain/interfaces/local_album.interface.dart';
+import 'package:immich_mobile/domain/interfaces/local_asset.interface.dart';
 import 'package:immich_mobile/domain/interfaces/log.interface.dart';
+import 'package:immich_mobile/domain/interfaces/storage.interface.dart';
 import 'package:immich_mobile/domain/interfaces/store.interface.dart';
 import 'package:immich_mobile/domain/interfaces/sync_api.interface.dart';
 import 'package:immich_mobile/domain/interfaces/sync_stream.interface.dart';
@@ -17,6 +20,12 @@ class MockDeviceAssetRepository extends Mock
     implements IDeviceAssetRepository {}
 
 class MockSyncStreamRepository extends Mock implements ISyncStreamRepository {}
+
+class MockLocalAlbumRepository extends Mock implements ILocalAlbumRepository {}
+
+class MockLocalAssetRepository extends Mock implements ILocalAssetRepository {}
+
+class MockStorageRepository extends Mock implements IStorageRepository {}
 
 // API Repos
 class MockUserApiRepository extends Mock implements IUserApiRepository {}
